Added a test that check balance of generated Rs.
It exhaustively generates matrix R for number of nodes up to 100 and
number of replicas up to 3 and checks that evaluation is zero.

Unfortunately, as of now this test fails because for few
combinations of parameters when number of nodes approaches
100. Because it turns out that the chance of generating slightly
unbalanced matrix R for these parameters is not negligible.