Fix Flow() for edges with demands.
When an edge with demand is added to the graph, it implicitly creates
two hidden edges.

For instance,

      Capacity: 10, Demand: 1
 A ---------------------------> B

becomes

          Capacity: 9
   A ---------------------------> B
   |                              ^
   | Capacity: 1                  | Capacity: 1
   v                              |
 DEMAND                        SUPPLY

Flow() is supposed to return the full flow through an edge (hiding the
details about any hidden edges). So it needs to add the flow through
the actual edge and the flow through the demand edges. The last
bit (calculating the flow through demand edges) worked incorrectly in
graphs with no feasible flow (graphs where not all demands are
satisfied). The old code took the minimum of the flows of the two
demand edges. If there's a feasible flow, both flows through the
demand edges will be the same. With no feasible flow, it's possible
that the demand edges will have different values if there are multiple
edges with demands from A or to B (what in-flows through SUPPLY->B may
out-flow through A->DEMAND created for a different vertex).

To address the issue this patch only looks at the flow through the
SUPPLY->B edge.

Change-Id: Id11c332614e2bbd4c546c691ee2348b8cc3d8d35