MB-50712: Introduce Vbmap greedy appproach
This patch adds a greedy optimization capability to vbmap. This approach
is a good deal simpler than the max flow approach that we currently use.
This simplicity allows us to fairly straight-forwardly introduce the idea of
comparing the vbmap we are generating to current vbmap and attempting to
minimize the distance (i.e. the total number of new replicas created)
between the two maps.

The greedy algorithm is actually 3 greedy assignment problems done
sequentially. These are
* slave assignment: assigning replication slave nodes for each node
* active vbucket assignment: assigning active vbuckets to nodes
* replica assignment: assigning replicas to nodes for each active
                      vbucket

Important side note: we should change our terminology from "slave" to
"secondary" at some point. Not in this patch, but probably soon.

All of the greedy assignment problems have the same high-level
structure:
* find the set of valid nodes that could be assigned
* from this set find the subset that minimizes some cost function
  (typically the cost is some count of the number of things a
  candidate node is already carrying)
* if a previous vbucket map is provided find the intersection of
  the least cost candidates with comparable nodes from the
  previous vbucket map
* choose a random node from the set of candidates

An important detail is that within the assignments if we don't like the
result of an assignment, we will set the current assignment aside and
work on other assignements until we find one that we like.

The best example of this is replica assignment (though we also do the same
thing in the slave assignment problem). Say we are assigning the first replica
for vb 101 and we find that there is no valid least cost node that
doesn't result in the creation of a new replica for this vbucket (when
compared with the previous vbucket map.) We set this vbucket aside and
try assigning the first replica for other vbuckets until we find one
that doesn't result in the creation of a new replica -- and we assign
that. The hope is that when we come back to vb 101 the costs will have
changed in such a way that we will then be able to assign a first
replica for it that doesn't result in the creation of a new replica.
This illustrates the essentially greedy nature of the algorithm.

This "exhaustive search" in attempting to find good replica nodes to
assign does give the algorithm a O(num-vbuckets^2) worst case runtime.
However, in practice the runtime is quick and stable -- typically
averaging around 0.2 s with worst case at around 0.5s for 1024 vbuckets.

Change-Id: I4efff1f8a46e08d099e475f5142425aee222a688
Reviewed-on: https://review.couchbase.org/c/vbmap/+/175352
Well-Formed: Restriction Checker
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>