MB-50712: Add the "rebalance move checker"
The rebalance move checker runs the following test:
* adds a single node to the cluster
* simulates running a rebalance

And determins the number of new replicas that end up getting
built as a result. The best case number of new replicas is:
    (num_vbuckets / N) * (num_replicas + 1)

where N is the number of nodes in the cluster before the
rebalance. The rationale is that as the cluster has increased from
N to N+1, 1/N of the data has to move from the existing cluster
to the new node for the new cluster to be balanced.

The rebalance move checker checks that the number of new replicas
built is less than 1.5 times the best case.

Change-Id: If8626f1796643f2783436a56730b918f68b800f7
Reviewed-on: https://review.couchbase.org/c/vbmap/+/175356
Well-Formed: Restriction Checker
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>